feat(copy): Enhance compression feedback during large transfers
- Update compressDirectory to return detailed results including
  number of files, compressed size, and time taken.
- Improve logging for compression process in copyToServer.

+++ tools/server-operations/src/utils/files.ts
@@ -40,5 +40,5 @@ export const fileCount = async (dir: string): Promise<number> => {
 };
 
 export const bytesToOthers = (bytes: number, unit: SizeUnit = 'M'): number => {
-  return parseInt((bytes / sizeUnits[unit]).toFixed(2));
+  return parseFloat((bytes / sizeUnits[unit]).toFixed(2));
 };
